+++ log_test.go
@@ -4,9 +4,12 @@
 package vai
 
 import (
+	"os"
+	"strings"
 	"testing"
 
 	"github.com/charmbracelet/log"
+	"github.com/muesli/termenv"
 	"github.com/stretchr/testify/require"
 )
 
@@ -23,4 +26,48 @@ func TestLogger(t *testing.T) {
 	SetLogLevel(log.DebugLevel)
 
 	require.Equal(t, log.DebugLevel, l.GetLevel())
+}
+
+func TestPrintScript(t *testing.T) {
+	testCases := []struct {
+		name     string
+		script   string
+		prefix   string
+		expected string
+	}{
+		{
+			name:     "simple eval",
+			script:   `h := "hello"`,
+			prefix:   ">",
+			expected: "> h := \"hello\"\n",
+		},
+		{
+			name:     "simple shell",
+			script:   "echo hello",
+			prefix:   "$",
+			expected: "$ echo hello\n",
+		},
+		{
+			name:     "multiline",
+			script:   "echo hello\necho world\n\necho !",
+			prefix:   "$",
+			expected: "$ echo hello\n$ echo world\n$ echo !\n",
+		},
+	}
+
+	var buf strings.Builder
+	logger.SetOutput(&buf)
+	_loggerColorProfile = termenv.Ascii
+	t.Cleanup(func() {
+		logger.SetOutput(os.Stderr)
+		_loggerColorProfile = 0
+	})
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			printScript(tc.prefix, tc.script)
+			require.Equal(t, tc.expected, buf.String())
+			buf.Reset()
+		})
+	}
 }
