+++ run_test.go
@@ -5,12 +5,9 @@ package vai
 
 import (
 	"context"
-	"os"
-	"strings"
 	"testing"
 	"time"
 
-	"github.com/muesli/termenv"
 	"github.com/noxsios/vai/storage"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/require"
@@ -116,50 +113,6 @@ func TestToEnvVar(t *testing.T) {
 			t.Parallel()
 			actual := toEnvVar(tc.s)
 			require.Equal(t, tc.expected, actual)
-		})
-	}
-}
-
-func TestPrintScript(t *testing.T) {
-	testCases := []struct {
-		name     string
-		script   string
-		prefix   string
-		expected string
-	}{
-		{
-			name:     "simple eval",
-			script:   `h := "hello"`,
-			prefix:   ">",
-			expected: "> h := \"hello\"\n",
-		},
-		{
-			name:     "simple shell",
-			script:   "echo hello",
-			prefix:   "$",
-			expected: "$ echo hello\n",
-		},
-		{
-			name:     "multiline",
-			script:   "echo hello\necho world\n\necho !",
-			prefix:   "$",
-			expected: "$ echo hello\n$ echo world\n$ echo !\n",
-		},
-	}
-
-	var buf strings.Builder
-	logger.SetOutput(&buf)
-	_loggerColorProfile = termenv.Ascii
-	t.Cleanup(func() {
-		logger.SetOutput(os.Stderr)
-		_loggerColorProfile = 0
-	})
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			printScript(tc.prefix, tc.script)
-			require.Equal(t, tc.expected, buf.String())
-			buf.Reset()
 		})
 	}
 }
